+++ collab-server/collab-error.ts
@@ -2,10 +2,11 @@ export class CollabError extends Error {
   from?: string;
   errorCode: number;
   body: any;
-  constructor(errorCode: number, body: any) {
-    super(body);
+  constructor(errorCode: number, message: any) {
+    super(message);
     this.errorCode = errorCode;
-    this.body = body;
+    this.body = message;
+
     // Error.captureStackTrace is a v8-specific method so not avilable on
     // Firefox or Safari
     if ((Error as any).captureStackTrace) {