+++ collab-server/index.ts
@@ -1,4 +1,13 @@
+import type {
+  PullEventResponse,
+  PullEvents,
+  GetDocument,
+  PushEvents,
+} from './types';
+
 export * from './instance';
 export * from './manager';
 export * from './collab-error';
 export * from './utils';
+export * from './parse-collab-response';
+export type { PullEventResponse, PullEvents, GetDocument, PushEvents };